Add MQ2Items.cpp, refactor Keyring handling
This started off as trying to remove the thread that was processing keyrings
as it can cause corruption or crashes.

Ended up refactoring keyrings to avoid using the UI and use item containers
directly from the character data.

Parameterized a lot of things on keyring type to avoid copy/paste of the
three/four different types of keyrings.

Fixed MaybeExactCompare

Added Spawns and Items modules.