Keyring TLOs use UI indices, add .Item to keyrings
Bring back the KeyRingWnd usage for the TLO only. The TLO wants indices
correlating to the order of the UI list.

Refreshing the KeyRingWindow is no synchronous and no longer requires a
state machine to perform. Normally, the keyring window is populated in
OnProcessFrame. Because CXWnd is mapped out and virtual, we can call
OnProcessFrame for the CKeyRingWnd. It also gates its update based on
the last timestamp and the current tab, so we temporarily set those, too.

Internally we are using indices into the item container. we map these
values to/from the ui when we deal with indices in the data type.

Added a .Item member to get the actual item. We're using the item
container indices so this is now easily possible.