Cleanup GetAllMercDesc
Moved out of MQ2Main.h since it exports stl types and so isn't safe for
use across dll boundaries.

Refactored to return a vector instead of taking a reference to a map.

Added a MQ2Utilities.h header which is private. May change this later as
we accumulate more private headers...