Refactor SpawnType: No more new/delete/memcpy
Datatypes will store a observed spawn. That observed spawn can be copied
and it will be nulled out automatically.

Added helpers to deal with creating and accessing these observed spawns.

Added Downcast feature to types: Allowing something like Target to downcast
to a Spawn so that inherited members can be evaluated without risk of
losing information.