Partial revert of #362
We can't return a nil when accessing data members prior to evaluation
because sometimes we don't know if they're valid or not.

The rules now are as follows:

If a mq data member is valid, then accessing its members will perform a
validity check. If a mq data member is invalid, then no validity check
is performed.

The result is that doing a tostring() on an un-evaluated data member will
return NULL if it is invalid.

Evaluating a data member by calling it with () will perform the validity
check and return nil if it is not valid.

Example:
   with no target: mq.TLO.Target.Body.Name will return a userdata object
   with tostring() value of "NULL". Can also compare against mq.null.

   mq.TLO.Target.Body() and mq.TLO.Target.Body.Name() will both return nil.