Restructure MapObjectMapLoc and MapLocTemplate organization
MapObjectMapLoc represents the Map representation of the object, a.k.a.
the map object. It no longer holds any relevant state of its own. These
instances are ephemeral and so any state must be held elsewhere. When
one of these is created, it receives a pointer back to the object that
created it only so that it can be informed when the map object is
destroyed.

MapLocTemplate represents the parameters to create the map object. It is
the "map loc". It is guaranteed to have a longer life span than the map
object. Any changes to the map loc will be propagated to the map object
as necessary. All bookkeeping and state management happens with the map
loc. When it is destroyed it automatically cleans up its map object.

No code other than map object management code ever needs to touch a map
object, and map management code never needs to touch a map loc.

- Also added some improvements to the imgui portion.