Fix unsafe_function in require hook bombing out and skipping destructor
The unsafe_function call is using some very unsafe internal mechanisms
which somehow skip stack unwinding when they fail, resulting in this
ScopedYieldDisabler not being destroyed.

The worst part is that it holds a reference count to the lua thread,
which prevents it from cleaning up. Basically, its leaking the script in
memory.

Change to a safe function call, which will swallow an error. So print the
error if one occurs.

Fixes #744.