Update MQ2GroupType.cpp
Expanded on the idea of Group.Cleric and added the remaining classes. Also added options Furthest, FurthestMelee, and FurthestPureCaster will return the 3D distance of the furthest player in the group of those types. To make it easier to figure out if group members are close enough for group type spells/effects. IE 2.0 clicks / Group heals.