Update MQ2GroupType.cpp
Got rid of the redundancy in the code. Can now just do a ${Group.ClassInGroup[<class_name>]} to return the spawn type of the class in the group or null if that class isn't in the group. Class name is a case insensitive compare. Changed furthest to just FurthestMember. Can be called with just FurthestMember, FurtherMember[purecaster], or FurtherestMember[melee] to get the non 3D distance to the furthest group member out depending on the criteria. Purecaster and melee are case insensitive as well.