Hack together some fixes for window inspector in login
EQ is using a *copy* of the CXWnd etc window hierarchy that differs from
the rest of the client. This makes full support for the window inspector
in login a little tricky...

Hacked together some basic support by just copy/pasting the CXWnd
viewer and editing out the parts that aren't working right.