added ability to save font selection between sessions
added ability to save and load a chosen font as default next Session

on larger resolution screens i use the font picker in the demo to change ImGui fonts, which gets tedious to do each session.

Checks MacroQuest.ini for a default font setting if found it will try to load that font when building out the font atlas, otherwise uses the ImGui defaults.

This brings the picker from the demo to the settings window and saves the chosen font to the ini for next load