Refactor how lua searches for scripts by path
When we find a script to run we also need to know where we found it.

This was simpler until I added multiple search paths, now we need to keep
track of which search path was used to find it, this way canonical name
can be calculated from it correctly.