Add support for %
Cherry-pick from upstream zep, plus other related changes:

Behaves the same way as in Vim.  % to go to the next match of the first
thing found on the line, or the previous thing found on the line, at any
further point into the document.  No support for a count.

Unit test added to confirm basic behavior.

Added a FindFirstCharOf command on the buffer, which returns the first
found char in a list of characters, either forward or backwards
(inclusive).

Co-authored-by: Chris Maughan <mornymorny@gmail.com>