Add experimental mode to copy loader to temp dir
This is disabled by default.

This behavior will copy the loader to the %TEMP% directory when it relaunches
itself. This is untested and experimental.

I will very likely change this completely.

Also changes the app to rely on the working directory instead of the
process path. To facilitate this, the working directory is updated at
launch.